Fix a rare crash when editing podcasts.
A podcast can be deleted while its feed is being refreshed. When this
happens the update transaction needs to check that the podcast row still
exists.

+++ app/src/main/java/com/studio4plus/homerplayer2/podcasts/usecases/UpdatePodcastEpisodes.kt
@@ -45,9 +45,11 @@ class UpdatePodcastEpisodes(
     ) {
         db.withTransaction {
             val feedUri = podcast.feedUri
-            podcastsDao.updatePodcast(feedUri, newTitle)
-            val oldBookIds = podcastsDao.updateEpisodes(feedUri, newEpisodes)
-            oldBookIds.forEach { audiobooksDao.deleteAudiobook(it) }
+            val updateCount = podcastsDao.updatePodcast(feedUri, newTitle)
+            if (updateCount > 0) { // Otherwise the podcast is no longer there.
+                val oldBookIds = podcastsDao.updateEpisodes(feedUri, newEpisodes)
+                oldBookIds.forEach { audiobooksDao.deleteAudiobook(it) }
+            }
         }
     }
 }
